shemas.pyについて
